feat: make Login grant work
Continuing on making what is currently documented in the README a reality, and
specifically in this change making the "Login" grant work. It's admittedly not
a GRANT in terms of PostgreSQL, but it seems reasonable to use the same API for
all "properties" of a user that allows them to do stuff.

We do avoid making changes where possible and avoid grabbing the exclusive
advisory lock if we don't need it, but it's not possible to see an existing
password. So, if a password is passed then we assume it has to be set.

Co-authored-by: Tash Boyse <57753415+nboyse@users.noreply.github.com>
Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Mohizur Khan <mohizurkhan@digital.trade.gov.uk>
Co-authored-by: Nick Ross <nick.ir.ross@gmail.com>
Co-authored-by: Josh Wong <166488409+joshwong-cs@users.noreply.github.com>