feat: ability to ignore tables in certain schemas - preserve team schema permissions
pg-sync-roles doesn't yet have the ability to handle all use cases, and
specifically certain situations of "team schemas" such as those in Data
Workspace, where many users should have access to everything in a schema, even
ones created in the future, after any call to sync_roles. Specifically, the
sync_roles function would revoke access to tables in team schemas.

To work around this, this change adds the ability to ignore all objects in
certain schemas, and so not revoke access to them.

Co-authored-by: Tash Boyse <57753415+nboyse@users.noreply.github.com>
Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Mohizur Khan <mohizurkhan@digital.trade.gov.uk>
Co-authored-by: Nick Ross <nick.ir.ross@gmail.com>
Co-authored-by: Josh Wong <166488409+joshwong-cs@users.noreply.github.com>