fix: ensure can revoke permissions when not owner
To control the permissions on an object in PostgreSQL, the role doing the
control needs to be owner on the object (or at least in PostgreSQL 16, be able
to SET ROLE to the role that's the owner).

This change makes the syncing user temporarily the owner of tables when we
REVOKE on them in order to do this.

There will probably be changes coming up doing this for when we GRANT

Co-authored-by: Tash Boyse <57753415+nboyse@users.noreply.github.com>
Co-authored-by: Michal Charemza <michal@charemza.name>
Co-authored-by: Mohizur Khan <mohizurkhan@digital.trade.gov.uk>
Co-authored-by: Josh Wong <166488409+joshwong-cs@users.noreply.github.com>