feat: revoke direct privileges on schemas
This I think has revealed some oddities when managing objects in schemas on
which the syncing user doesn't have USAGE permission, which is revoked
automatically when a schema's ownership is transferred to another user.
Specifically, the syncing user then can't even see tables in the schema.

The best-effort change work around here is to grant itself the owner role of
the schemas to try to get USAGE on them.

I am starting to suspect that ownership of schemas is a bit tricky to be
delegated. Granting USAGE and CREATE is probably the way to go to allow roles
to create things in them.