feat: allow permissions on an object to be granted to a role directly
This adds the ability to bypass the intermediate role system that pg-sync-roles
was created with. This is to reduce the numbers of roles that users are
granted.

While the intermediate role system addressed some issues ("row is too big"
errors with high numbers of users granted access to tables), it now seems to
have introduced another issue. For users with high numbers of roles granted,
they can take a long, > 90 second, time to connect to the database.

In our usage of pg-sync-roles, all access to tables is not granted on an
individual basis, but a large proportion is granted based on a membership of a
low number of real-world groups that the users are members of. So we can make
our own intermediate database roles that match these groups, make the users
members of these roles, and give the roles direct access to the database
objects, without the per-database-object-grant-type intermediate roles, and so
greatly reducing to the total numbers of roles that a user is
directly/indirectly a member of compared to using the intermediate roles.

The one part of this that wasn't possible with pg-sync-roles is the ability for
roles to have direct access to database objects, specifically just schema USAGE
and table SELECT, which make up the vast majority of the intermediate roles in
our case.

Co-authored-by: Peter Woodcock <peter.woodcock@digital.trade.gov.uk>
Co-authored-by: Michal Charemza <michal@charemza.name>