refactor: use correct variable name in function
This changes one of the internal functions to use the variable passed to it,
rather than the one in the surrounding scope.

It shouldn't change any functionality, the only use of the variable is to cast
it to a bool, and the variables are always the same in terms of their
truthyness.