fix: incorrectly not granting when only a direct table change
This fixes an issue where if the only change is the addition of a direct table
permission, the permission wouldn't be granted, because the sync_roles function
would incorrectly identifty there is nothing to do.

If there were other changes to be made, then permissions would be still be
granted correctly.