feat: allow regex table names in TableSelect (#144)
This extends TableSelect by allowing the table name to be a regular
expression, which results in permissions granted to all tables in the
schema whose name matches the regular expression.

This addresses https://github.com/uktrade/pg-sync-roles/issues/143 that
requested being able to choose all tables in a schema. This is possible
via this change by using `TableSelect('my_schema', re.compile('.*'))`