perf: remove unnecessary queries after obtaining lock
This removes re-checking if tables/schemas/database exist after obtaining lock.

This might technically very slightly increase the chance of errors if say a
schema is removed during the run of sync_roles, but the lock is not for
wrapping the removal of tables/schemas(/database...), and so we are already
vulnerable to that in various ways I suspect. I am pro to think about how to
sort that properly later, or leave that as out of scope and a possible risk
that calling code has to deal with.